[Refacot] change youtube service dir
