+++ src/services/playlist/youtube/track.js
@@ -1,44 +1,8 @@
 const axios = require("axios");
 
-const { cacheService } = require("../database");
+const { cacheService } = require("../../database");
 
-const searchList = async (token) => {
-  try {
-    const params = {
-      part: "snippet",
-      maxResults: 5,
-      mine: true,
-      pageToken: "",
-    };
-    const options = {
-      method: "GET",
-      url: "https://www.googleapis.com/youtube/v3/playlists",
-      headers: {
-        authorization: `Bearer ${token}`,
-      },
-      params,
-    };
-
-    const playLists = [];
-    do {
-      const response = await axios(options);
-      const { data } = response;
-
-      data.items.forEach((item) => {
-        playLists.push(parsePlayList(item));
-      });
-
-      params.pageToken = data.nextPageToken;
-    } while (params.pageToken);
-
-    return { playLists };
-  } catch (error) {
-    console.error(error);
-    throw new Error(error);
-  }
-};
-
-const getPlayListItem = async (id, token) => {
+const getId = async (id, token) => {
   try {
     const params = {
       part: "contentDetails",
@@ -72,7 +36,7 @@ const getPlayListItem = async (id, token) => {
   }
 };
 
-const getTrackInfo = async (id, token) => {
+const getInfo = async (id, token) => {
   try {
     const params = {
       part: "snippet",
@@ -111,41 +75,7 @@ const getTrackInfo = async (id, token) => {
   }
 };
 
-const createPlayList = async (playLists, token) => {
-  try {
-    const params = {
-      part: "snippet",
-    };
-    const data = {
-      snippet: {
-        title: "",
-      },
-    };
-    const options = {
-      method: "POST",
-      url: "https://www.googleapis.com/youtube/v3/playlists",
-      headers: {
-        authorization: `Bearer ${token}`,
-      },
-      params,
-      data,
-    };
-
-    const newPlayLists = [];
-    for (const playList of playLists) {
-      data.snippet.title = playList.title;
-      const response = await axios(options);
-      newPlayLists.push(response.data.id);
-    }
-
-    return newPlayLists;
-  } catch (error) {
-    console.error(error);
-    throw new Error(error);
-  }
-};
-
-const insertTracks = async (id, tracks, token) => {
+const create = async (id, tracks, token) => {
   try {
     const params = {
       part: "snippet",
@@ -180,41 +110,9 @@ const insertTracks = async (id, tracks, token) => {
   }
 };
 
-const splitArray50 = (array) => {
-  let start = 0;
-  let end = 50;
-  const result = [];
-  while (start < array.length) {
-    result.push(array.slice(start, end));
-    start = end;
-    end += 50;
-  }
-  return result;
-};
+module.exports = { getId, getInfo, create };
 
-module.exports = {
-  searchList,
-  getPlayListItem,
-  getTrackInfo,
-  createPlayList,
-  insertTracks,
-  splitArray50,
-};
-
-//not exports//
-const parsePlayList = (playList) => {
-  return {
-    id: playList.id,
-    title: playList.snippet.title,
-    thumbnail: playList.snippet.thumbnails.default,
-    description: playList.snippet.description,
-    owner: {
-      name: playList.snippet.channelTitle,
-      id: playList.snippet.channelId,
-    },
-  };
-};
-
+//not exports
 const parseTrackItem = (trackId) => {
   return trackId.contentDetails.videoId;
 };
